Proper handling of non breaking space
The bug I'm fixing with this is parsing of "<em>feelings:&nbsp;</em>". Non breaking space either needs to be removed, or preserved as HTML entitity. They are converted to whitespace instead.

Since it is a specific none breaking character, I think it's better to preserve it as that is the intention. However it could be added as a configuration option.