Add treat_children to pre converter
Adds treat_children to the pre converter. This resulted in one special
case where a code block directly nested in a pre block as the only child
would have unnecessary backticks around the whole string, so I have
added one line to chop them off in that case.

I also updated the node_for function in spec_helper so that it parses as
HTML