Add support for JRuby
It seems the JRuby version of Nokogiri explicitly includes an empty head
element when parsing a HTML fragment using `Nokogiri::HTML(fragment)`,
whereas the MRI version does not. While it would be reasonable for
Nokogiri to behave the same way on all platforms, it's not obvious
whether that would be to remove head or to always include it, and given
that the simple workaround to ignore head elements is sufficient to
support both, it seems reasonable to not wait until Nokogiri is updated.

In order to avoid similar issues creeping up in the future, this adds a
JRuby target to Travis, and patches the test suite to be compatible. In
addition to handling the head element in `node_for`, it meant replacing
redcarpet that requires a C-extension with pure Ruby kramdown instead.

While I suppose kramdown could be used for all targets, the current
implementation keeps using redcarpet unless we are using JRuby, which
effectively means that the test-suite ensure compatiblity with both.