Fixes unintentional newline characters within lists with paragraphs
This commit aims to fix a problem when parsing markup similar to:

```ruby
[8] pry(main)> ReverseMarkdown.convert("<ul><li> <p>a</p></li></ul>")
=> "-  \n\na\n\n"
```

Note that the list item has newlines between the markdown list item and
the actual text.

If you remove the space between the list item tag and the paragraph tag:

```ruby
[7] pry(main)> ReverseMarkdown.convert("<ul><li><p>a</p></li></ul>")
=> "- a\n\n"
```

Reverse markdown appears to assume what believe to be the intended
behaviour. There was already a test in place to account for this situation,
but it was `xit`'ed.

The proposed patch is to ask Nokogiri to find the first child element of
the list, instead of the first child (which might include stuff like a
newline).